+++ src/app/app.module.ts
@@ -8,7 +8,8 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 import { NouisliderModule } from 'ng2-nouislider';
 import { ClickOutsideModule } from 'ng4-click-outside';
-import { Ng2CompleterModule } from 'ng2-completer';
+import { TreeModule } from 'angular-tree-component';
+import { MultiselectDropdownModule } from 'angular-2-dropdown-multiselect';
 
 // ANGULAR MATERIAL MODULES
 import { MatSlideToggleModule, MatButtonModule, MatCheckboxModule, MatExpansionModule, MatProgressSpinnerModule } from '@angular/material';
@@ -26,7 +27,7 @@ import { TitleCasePipe } from '@angular/common';
 import { ProjectsFilterPipe } from './_shared/pipes/projects-filter.pipe';
 
 // SERVICES
-import { CacheService, ClickTrackingService, CookiesService,
+import { BomService, CacheService, ClickTrackingService, CookiesService,
   ToolsService, UserResolverService, WebsocketService, RoutingHistoryService, ExcelExportService } from './_shared/services/_index';
 import { AuthService } from './_shared/services/auth.service';
 import { ApiDataAuthService, ApiDataClickTrackingService, ApiDataEmailService, ApiDataEmployeeService,
@@ -40,11 +41,11 @@ import { UnsavedChangesGuard } from './_shared/guards/unsaved-changes.guard';
 import { BrowserGuard } from './_shared/guards/browser.guard';
 import { PermissionsGuard } from './_shared/guards/permissions.guard';
 import { FteEntryGuard } from './fte-entry/employee/fte-entry.guard';
+import { FteTeamEntryGuard } from './fte-entry/team/fte-entry.guard';
 
 // CHARTS
 import 'hammerjs';
 import 'chartjs-plugin-zoom';
-import { TreeModule } from 'ng2-tree';
 
 // COMPONENTS
 import { AppComponent } from './app.component';
@@ -61,7 +62,7 @@ import { EmployeesReportsComponent } from './reports/employees/employees.compone
 import { ProfileModalComponent } from './modals/profile-modal/profile-modal.component';
 import { ProjectsModalComponent } from './modals/projects-modal/projects-modal.component';
 import { TestComponent } from './test/test.component';
-import { OrgDropdownComponent } from './reports/team-fte-summary/org-dropdown/org-dropdown.component';
+import { OrgDropdownComponent } from './reports/employees/org-dropdown/org-dropdown.component';
 import { ProjectsEditModalComponent } from './modals/projects-edit-modal/projects-edit-modal.component';
 import { ProjectsCreateModalComponent } from './modals/projects-create-modal/projects-create-modal.component';
 import { ChartsModule } from 'ng2-charts';
@@ -91,7 +92,10 @@ import { BomViewerComponent } from './bom/bom-viewer/bom-viewer.component';
 import { SearchProjectsComponent } from './projects/search-projects/search-projects.component';
 import { MyProjectsComponent } from './projects/my-projects/my-projects.component';
 import { ProjectRequestsComponent } from './projects/project-requests/project-requests.component';
+import { TeamRolesComponent } from './setups/team-roles/team-roles.component';
 import { DisplayProjectComponent } from './projects/display-project/display-project.component';
+import { UnitTestComponent } from './unit-test/unit-test.component';
+import { MatplanComponent } from './mat-plan/matplan/matplan.component';
 
 
 @NgModule({
@@ -148,7 +152,10 @@ import { DisplayProjectComponent } from './projects/display-project/display-proj
     BomViewerComponent,
     ProjectRequestsComponent,
     SearchProjectsComponent,
-    DisplayProjectComponent
+    TeamRolesComponent,
+    DisplayProjectComponent,
+    UnitTestComponent,
+    MatplanComponent
   ],
   imports: [
     AppRoutingModule,
@@ -165,8 +172,8 @@ import { DisplayProjectComponent } from './projects/display-project/display-proj
     NouisliderModule,
     ClickOutsideModule,
     ChartsModule,
-    Ng2CompleterModule,
-    TreeModule
+    TreeModule.forRoot(),
+    MultiselectDropdownModule
   ],
   providers: [
     ApiDataAuthService,
@@ -185,6 +192,7 @@ import { DisplayProjectComponent } from './projects/display-project/display-proj
     ApiDataPartService,
     ApiDataAnalyticsService,
     ApiDataBomService,
+    BomService,
     CacheService,
     AuthService,
     AuthGuardService,
@@ -197,6 +205,7 @@ import { DisplayProjectComponent } from './projects/display-project/display-proj
     BrowserGuard,
     PermissionsGuard,
     FteEntryGuard,
+    FteTeamEntryGuard,
     TitleCasePipe,
     RoutingHistoryService,
     ExcelExportService