+++ src/app/fte-entry/employee/fte-entry.component.ts
@@ -1,9 +1,8 @@
-import { Component, OnInit, ViewChild, AfterViewInit, OnDestroy, Input, ChangeDetectorRef } from '@angular/core';
+import { Component, OnInit, ViewChild, OnDestroy, ChangeDetectorRef } from '@angular/core';
 import { FormGroup, FormArray, FormControl, Validators, FormBuilder } from '@angular/forms';
-import { trigger, state, style, transition, animate, keyframes, group } from '@angular/animations';
+import { trigger, state, style, transition, animate } from '@angular/animations';
 import { DecimalPipe } from '@angular/common';
 import { HostListener } from '@angular/core';
-import { NouisliderModule } from 'ng2-nouislider';
 import { Subscription } from 'rxjs/Subscription';
 import { Observable } from 'rxjs/Observable';
 
@@ -15,7 +14,6 @@ import { ComponentCanDeactivate } from '../../_shared/guards/unsaved-changes.gua
 import { UserFTEs, AllocationsArray} from './fte-model';
 import { utils, write, WorkBook } from 'xlsx';
 import { saveAs } from 'file-saver';
-import { JAN } from '@angular/material';
 
 import { ProjectsCreateModalComponent } from '../../modals/projects-create-modal/projects-create-modal.component';
 
@@ -432,8 +430,8 @@ export class FteEntryEmployeeComponent implements OnInit, OnDestroy, ComponentCa
   onTestFormClick() {
     // console.log('form object (this.form):');
     // console.log(this.FTEFormGroup);
-    // console.log('form data (this.form.value.FTEFormArray):');
-    // console.log(this.FTEFormGroup.value.FTEFormArray);
+    console.log('form data (this.form.value.FTEFormArray):');
+    console.log(this.FTEFormGroup.value.FTEFormArray);
     // console.log('fte-project-visible array');
     // console.log(this.fteProjectVisible);
   }
@@ -456,15 +454,19 @@ export class FteEntryEmployeeComponent implements OnInit, OnDestroy, ComponentCa
 
     // validate that no projects are empty
     const fteData = this.FTEFormGroup.value.FTEFormArray;
+    const emptyProjects = [];
     const noProjectsEmpty = fteData.every( project => {
       // check if every editable value for a given project is null
-      const isProjectEmpty = project.slice(firstEditableMonth).every( entry => {
+      const isProjectEmpty = project.every( entry => {
         return entry.fte === null;
       });
+      if (isProjectEmpty) {
+        emptyProjects.push(project[0].projectName);
+      }
       return !isProjectEmpty;
     });
 
-    // only save if all quarters are valid
+    // only save if all quarters are valid and no projects are empty
     if (currentQuarterValid && futureQuartersValid && noProjectsEmpty) {
       const t0 = performance.now();
 
@@ -512,18 +514,9 @@ export class FteEntryEmployeeComponent implements OnInit, OnDestroy, ComponentCa
 
     } else if (!noProjectsEmpty) {
       // if the save was disallowed because there are empty projects, let the user know which ones
-      const invalidProjects = [];
-      fteData.forEach( project => {
-        const isProjectEmpty = project.slice(firstEditableMonth).every( entry => {
-          return entry.fte === null;
-        });
-        if (isProjectEmpty) {
-          invalidProjects.push(` ${project[0].projectName}`);
-        }
-      });
-      const invalidProjectsString = invalidProjects.toString();
+      // const invalidProjectsString = ;
       this.cacheService.raiseToast('error', `All projects in your table must have FTE entries.
-      Please add FTE entries for the following projects:${invalidProjectsString} and try again`);
+      Please add FTE entries for the following projects: ${emptyProjects.toString()}`);
     } else if (!currentQuarterValid || !futureQuartersValid) {
       this.cacheService.raiseToast('error', `FTE totals in each month cannot exceed 100%.`);
     } else {
@@ -786,68 +779,73 @@ export class FteEntryEmployeeComponent implements OnInit, OnDestroy, ComponentCa
   }
 
   onTrashClick(index: number) {
-    // console.log('user clicked to delete project index ' + index);
-    const FTEFormArray = <FormArray>this.FTEFormGroup.controls.FTEFormArray;
-    const deletedProject: any = FTEFormArray.controls[index];
-    // emit confirmation modal after they click delete button
-    this.cacheService.confirmModalData.emit(
-      {
-        title: 'Confirm Deletion',
-        message: `Are you sure you want to permanently delete all of your FTE values for project ${deletedProject.projectName}?`,
-        iconClass: 'fa-exclamation-triangle',
-        iconColor: 'rgb(193, 193, 27)',
-        closeButton: true,
-        allowOutsideClickDismiss: false,
-        allowEscKeyDismiss: false,
-        buttons: [
-          {
-            text: 'Yes',
-            bsClass: 'btn-success',
-            emit: true
-          },
-          {
-            text: 'Cancel',
-            bsClass: 'btn-secondary',
-            emit: false
-          }
-        ]
-      }
-    );
-
-    const deleteModalSubscription = this.cacheService.confirmModalResponse.subscribe( res => {
-      if (res) {
-        // if they click ok, grab the deleted project info and exec db call to delete
-        const toBeDeleted = {
-          projectID: deletedProject.projectID,
-          projectName: deletedProject.projectName,
-          newlyAdded: deletedProject.newlyAdded
-        };
-
-        const deleteActionSubscription = this.apiDataFteService.destroyUserProject(toBeDeleted.projectID).subscribe(
-          deleteResponse => {
-            // only delete from the projectemployeerole table if user is deleting a non-newlyAdded project
-            if (!toBeDeleted.newlyAdded) {
-              this.deleteProjectEmployeeRole(toBeDeleted);
+    if (!this.fteProjectDeletable[index]) {
+      // if project is not deletable, raise an error toast
+      this.cacheService.raiseToast('error', 'Projects with historic FTE values cannot be deleted.  Hide them using the slider.');
+    } else {
+      // show a confirmation modal before deleting
+      const FTEFormArray = <FormArray>this.FTEFormGroup.controls.FTEFormArray;
+      const deletedProject: any = FTEFormArray.controls[index];
+      // emit confirmation modal after they click delete button
+      this.cacheService.confirmModalData.emit(
+        {
+          title: 'Confirm Deletion',
+          message: `Are you sure you want to permanently delete all of your FTE values for project ${deletedProject.projectName}?`,
+          iconClass: 'fa-exclamation-triangle',
+          iconColor: 'rgb(193, 193, 27)',
+          closeButton: true,
+          allowOutsideClickDismiss: false,
+          allowEscKeyDismiss: false,
+          buttons: [
+            {
+              text: 'Yes',
+              bsClass: 'btn-success',
+              emit: true
+            },
+            {
+              text: 'Cancel',
+              bsClass: 'btn-secondary',
+              emit: false
             }
-            this.fteProjectVisible.splice(index, 1);
-            this.fteProjectDeletable.splice(index, 1);
-            FTEFormArray.controls.splice(index, 1);
-            this.updateMonthlyTotals();
-            this.setMonthlyTotalsBorder();
-            // console.log('stuff was updated');
-            this.cacheService.raiseToast('success', deleteResponse.message);
-            deleteActionSubscription.unsubscribe();
-          },
-          deleteErr => {
-            this.cacheService.raiseToast('warning', `${deleteErr.status}: ${deleteErr.statusText}`);
-            deleteActionSubscription.unsubscribe();
-          }
-        );
-      } else {
-        // console.log('delete aborted');
-      }
-      deleteModalSubscription.unsubscribe();
-    });
+          ]
+        }
+      );
+
+      const deleteModalSubscription = this.cacheService.confirmModalResponse.subscribe( res => {
+        if (res) {
+          // if they click ok, grab the deleted project info and exec db call to delete
+          const toBeDeleted = {
+            projectID: deletedProject.projectID,
+            projectName: deletedProject.projectName,
+            newlyAdded: deletedProject.newlyAdded
+          };
+
+          const deleteActionSubscription = this.apiDataFteService.destroyUserProject(toBeDeleted.projectID).subscribe(
+            deleteResponse => {
+              // only delete from the projectemployeerole table if user is deleting a non-newlyAdded project
+              if (!toBeDeleted.newlyAdded) {
+                this.deleteProjectEmployeeRole(toBeDeleted);
+              }
+              this.fteProjectVisible.splice(index, 1);
+              this.fteProjectDeletable.splice(index, 1);
+              FTEFormArray.controls.splice(index, 1);
+              this.updateMonthlyTotals();
+              this.setMonthlyTotalsBorder();
+              // console.log('stuff was updated');
+              this.cacheService.raiseToast('success', deleteResponse.message);
+              deleteActionSubscription.unsubscribe();
+            },
+            deleteErr => {
+              this.cacheService.raiseToast('warning', `${deleteErr.status}: ${deleteErr.statusText}`);
+              deleteActionSubscription.unsubscribe();
+            }
+          );
+        } else {
+          // console.log('delete aborted');
+        }
+        deleteModalSubscription.unsubscribe();
+      });
+    }
   }
 
   onResetClick() {