+++ src/app/unit-test/unit-test.component.ts
@@ -1,0 +1,42 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-unit-test',
+  templateUrl: './unit-test.component.html',
+  styleUrls: ['./unit-test.component.css']
+})
+export class UnitTestComponent implements OnInit {
+
+  title: string;
+  message: string;
+  value: number;
+
+  constructor() {
+
+    this.title = 'unit-test works!';
+    this.value = 0;
+
+   }
+
+  ngOnInit() {
+  }
+
+  increment() {
+    if (this.value < 15) {
+      this.value += 1;
+      this.message = '';
+    } else {
+      this.message = 'Maximum reached!';
+    }
+  }
+
+  decrement() {
+    if (this.value > 0) {
+      this.value -= 1;
+      this.message = '';
+    } else {
+      this.message = 'Minimum reached!';
+    }
+  }
+
+}